refactor delete project functionality and test cases
