+++ src/test/kotlin/data/repository/project/DeleteProjectRepositoryTest.kt
@@ -1,6 +1,7 @@
 package data.repository.project
 
 import data.datasources.CsvDataSource
+import data.datasources.projectDataSource.ProjectDataSourceInterface
 import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
@@ -19,14 +20,14 @@ import java.util.UUID
 import java.util.UUID.randomUUID
 
 class DeleteProjectRepositoryTest {
-    lateinit var dataSource: CsvDataSource<Project>
+    lateinit var dataSource: ProjectDataSourceInterface
     lateinit var repository: ProjectRepositoryImpl
     lateinit var dummyId: UUID
 
     @BeforeEach
     fun setUp() {
-        dataSource = mockk()
-        repository = ProjectRepositoryImpl(mockk())
+        dataSource = mockk(relaxed = true)
+        repository = ProjectRepositoryImpl(dataSource)
         dummyId = randomUUID()
 
     }
@@ -34,47 +35,12 @@ class DeleteProjectRepositoryTest {
     @Test
     fun `should delete project successfully when project is already exist`() {
         // given
-        every {dataSource.getAll()} returns someProjects
-        every { dataSource.saveAll(any()) } just runs
+        every { dataSource.deleteProject(someProjects[0].id) } returns true
         assertDoesNotThrow {
             repository.deleteProject(someProjects[0].id)
 
         }
     }
-
-
-    @Test
-    fun `should call getAll and saveAll function form datasource  when trying to delete project`(){
-        every {dataSource.getAll()} returns someProjects
-        every { dataSource.saveAll(any()) } just runs
-        repository.deleteProject(someProjects[0].id)
-
-        verify(exactly = 1) { dataSource.getAll() }
-        verify(exactly = 1) {  dataSource.saveAll(any()) }
-
-    }
-
-    @Test
-    fun `should throw not found exception when deleting project not exist`() {
-        every {dataSource.getAll()} returns someProjects
-        every { dataSource.saveAll(any()) } just runs
-        // when & then
-        assertThrows<ProjectNotFoundException> {
-            repository.deleteProject(dummyId)
-        }
-
-    }
-    @Test
-    fun `should throw empty data exception when deleting project and there are not any project yet`() {
-        every {dataSource.getAll()} returns emptyList()
-        every { dataSource.saveAll(any()) } just runs
-        // when & then
-        assertThrows<EmptyDataException> {
-            repository.deleteProject(dummyId)
-        }
-
-    }
-
 
 
 }
