+++ src/test/kotlin/ui/project/DeleteProjectUiControllerTest.kt
@@ -1,18 +1,12 @@
-package ui
+package ui.project
 
 import io.mockk.every
-import io.mockk.just
 import io.mockk.mockk
-import io.mockk.runs
 import io.mockk.verify
 import logic.usecases.project.DeleteProjectUseCase
-import logic.usecases.task.EditTaskUseCase
-import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import ui.console.ConsoleIO
 import ui.console.ConsoleIOImpl
-import java.io.ByteArrayInputStream
 import java.util.UUID
 
 class DeleteProjectUiControllerTest {
@@ -30,7 +24,7 @@ class DeleteProjectUiControllerTest {
     @Test
     fun `should call delete project usecase when try to delete project`() {
         val projectId = UUID.randomUUID()
-        every { uiController.consoleIO.readFromUser()} returns projectId.toString()
+        every { uiController.consoleIO.readFromUser() } returns projectId.toString()
         uiController.execute()
 
 
@@ -40,9 +34,18 @@ class DeleteProjectUiControllerTest {
     @Test
     fun `should print success message to user when deleting project complete successfully`() {
         val projectId = UUID.randomUUID()
-        every { uiController.consoleIO.readFromUser()} returns projectId.toString()
-        every { useCase.deleteProject(projectId.toString())} returns true
+        every { uiController.consoleIO.readFromUser() } returns projectId.toString()
+        every { useCase.deleteProject(projectId.toString()) } returns true
         uiController.execute()
         verify { consoleIO.println("Project deleted successfully") }
+    }
+
+    @Test
+    fun `should print failure message to user when deleting project does not complete successfully`() {
+        val projectId = UUID.randomUUID()
+        every { uiController.consoleIO.readFromUser() } returns projectId.toString()
+        every { useCase.deleteProject(projectId.toString()) } returns false
+        uiController.execute()
+        verify { consoleIO.println("Failed deleting project")}
     }
 }
